#vfb38 Filter Menu component
